requirements.txt
langchain==0.0.153
openai==0.27.2
python-dotenv==1.0.0
google_search_results==2.4.2
pytest==7.3.1